sql_analyst_agent/requirements.txt
groq>=0.4.0
python-dotenv>=1.0.0
pandas>=2.0.0
tabulate>=0.9.0
scikit-learn>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
faker>=22.0.0
colorama>=0.4.6
